C2.8.2. Base for Chess app.
