+++ chess/src/main/java/en/firuges/black/PawnBlack.java
@@ -1,7 +1,7 @@
-package enUs.firuges.black;
+package en.firuges.black;
 
-import enUs.firuges.Cell;
-import enUs.firuges.Figure;
+import en.firuges.Cell;
+import en.firuges.Figure;
 
 public class PawnBlack implements Figure {
     private final Cell position;
@@ -11,7 +11,7 @@ public class PawnBlack implements Figure {
     }
 
     @Override
-    public Cell final position() {
+    public Cell position() {
         return this.position;
     }
 
@@ -19,7 +19,7 @@ public class PawnBlack implements Figure {
     public Cell[] way(Cell source, Cell dest) {
         Cell[] steps = new Cell[0];
         if (source.y == dest.y + 1 && source.x == dest.x) {
-            steps = new Cell[] { dest };
+            steps = new Cell[] {dest};
         }
         return steps;
     }
