+++ chess/src/main/java/en/firuges/black/QeenBlack.java
@@ -1,7 +1,7 @@
-package enUs.firuges.black;
+package en.firuges.black;
 
-import enUs.firuges.Cell;
-import enUs.firuges.Figure;
+import en.firuges.Cell;
+import en.firuges.Figure;
 
 public class QeenBlack implements Figure {
     private final Cell position;
@@ -17,7 +17,7 @@ public class QeenBlack implements Figure {
 
     @Override
     public Cell[] way(Cell source, Cell dest) {
-        return new Cell[] { dest };
+        return new Cell[] {dest};
     }
 
     @Override
