+++ chess/src/main/java/en/firuges/white/PawnWhite.java
@@ -1,12 +1,12 @@
-package enUs.firuges.white;
+package en.firuges.white;
 
-import enUs.firuges.Cell;
-import enUs.firuges.Figure;
+import en.firuges.Cell;
+import en.firuges.Figure;
 
-public class KingWhite implements Figure {
+public class PawnWhite implements Figure {
     private final Cell position;
 
-    public KingWhite(final Cell position) {
+    public PawnWhite(final Cell position) {
         this.position = position;
     }
 
@@ -17,11 +17,11 @@ public class KingWhite implements Figure {
 
     @Override
     public Cell[] way(Cell source, Cell dest) {
-        return new Cell[] { dest };
+        return new Cell[] {dest};
     }
 
     @Override
     public Figure copy(Cell dest) {
-        return new KingWhite(dest);
+        return new PawnWhite(dest);
     }
 }
