+++ chess/src/main/java/en/firuges/white/QeenWhite.java
@@ -1,12 +1,12 @@
-package enUs.firuges.black;
+package en.firuges.white;
 
-import enUs.firuges.Cell;
-import enUs.firuges.Figure;
+import en.firuges.Cell;
+import en.firuges.Figure;
 
-public class KnightBlack implements Figure {
+public class QeenWhite implements Figure {
     private final Cell position;
 
-    public KnightBlack(final Cell position) {
+    public QeenWhite(final Cell position) {
         this.position = position;
     }
 
@@ -17,11 +17,11 @@ public class KnightBlack implements Figure {
 
     @Override
     public Cell[] way(Cell source, Cell dest) {
-        return new Cell[] { dest };
+        return new Cell[] {dest};
     }
 
     @Override
     public Figure copy(Cell dest) {
-        return new KnightBlack(dest);
+        return new QeenWhite(dest);
     }
 }
